+++ src/services/goals.test.js
@@ -6,35 +6,21 @@ import db, {
   Grant,
   Goal,
   GoalTemplate,
-  GoalTemplateFieldPrompt,
-  GoalFieldResponse,
-  GoalResource,
   Objective,
-  ObjectiveResource,
-  ObjectiveFile,
-  ObjectiveTopic,
   ActivityReportGoal,
-  ActivityReportObjective,
-  Resource,
-  Topic,
-  File,
   ActivityReport,
   ActivityRecipient,
   User,
 } from '../models';
 import {
-  mergeGoals,
   reduceObjectives,
   reduceObjectivesForActivityReport,
   determineMergeGoalStatus,
-  getGoalsForReport,
   createMultiRecipientGoalsFromAdmin,
 } from './goals';
-import { createReport, destroyReport, createGoalTemplate } from '../testUtils';
 import {
   OBJECTIVE_STATUS,
   AUTOMATIC_CREATION,
-  FILE_STATUSES,
   GOAL_STATUS,
 } from '../constants';
 import { setFieldPromptsForCuratedTemplate } from './goalTemplates';
@@ -318,538 +304,6 @@ describe('Goals DB service', () => {
         GOAL_STATUS.DRAFT,
       ]);
       expect(status).toBe(GOAL_STATUS.DRAFT);
-    });
-  });
-
-  describe('mergeGoals', () => {
-    let recipient;
-    let grantOne;
-    let grantTwo;
-    let grantThree;
-    let report;
-    let template;
-    let topic;
-    let goalOne;
-    let goalTwo;
-    let goalThree;
-
-    let objectiveOneForGoalOne;
-
-    const fileName = faker.system.fileName();
-    const resourceUrl = faker.internet.url();
-
-    let mergedGoals;
-    let mergedGoalIds;
-
-    let dummyGoal;
-    const dummyGoalName = `dummy goal ${faker.animal.cetacean()} ${faker.datatype.string()}`;
-
-    beforeAll(async () => {
-      // Recipient.
-      recipient = await Recipient.create({
-        id: faker.datatype.number(),
-        name: faker.name.firstName(),
-        startDate: new Date(),
-        endDate: new Date(),
-      });
-
-      // Grant.
-      grantOne = await Grant.create({
-        id: faker.datatype.number(),
-        number: faker.datatype.string(),
-        recipientId: recipient.id,
-        regionId: 1,
-        startDate: new Date(),
-        endDate: new Date(),
-        status: 'Active',
-      });
-
-      grantTwo = await Grant.create({
-        id: faker.datatype.number({ min: grantOne.id + 1 }),
-        number: faker.datatype.string(),
-        recipientId: recipient.id,
-        regionId: 1,
-        startDate: new Date(),
-        endDate: new Date(),
-        status: 'Inactive',
-      });
-
-      grantThree = await Grant.create({
-        id: faker.datatype.number(),
-        number: faker.datatype.string(),
-        recipientId: recipient.id,
-        regionId: 1,
-        startDate: new Date(),
-        endDate: new Date(),
-        status: 'Active',
-        oldGrantId: grantTwo.id,
-      });
-
-      template = await createGoalTemplate();
-
-      const promptTitle = faker.datatype.string(255);
-
-      const prompt = await GoalTemplateFieldPrompt.create({
-        goalTemplateId: template.id,
-        ordinal: 1,
-        title: promptTitle,
-        prompt: promptTitle,
-        hint: '',
-        options: ['option 1', 'option 2', 'option 3'],
-        fieldType: 'multiselect',
-        validations: { required: 'Select a root cause', rules: [{ name: 'maxSelections', value: 2, message: 'You can only select 2 options' }] },
-      });
-
-      goalOne = await Goal.create({
-        name: `Selected goal 1${faker.animal.dog()}`,
-        status: GOAL_STATUS.IN_PROGRESS,
-        endDate: null,
-        isFromSmartsheetTtaPlan: false,
-        onApprovedAR: false,
-        grantId: grantOne.id,
-        createdVia: 'rtr',
-      });
-
-      goalTwo = await Goal.create({
-        name: `Selected goal 2${faker.animal.dog()}`,
-        status: GOAL_STATUS.NOT_STARTED,
-        endDate: null,
-        isFromSmartsheetTtaPlan: false,
-        onApprovedAR: false,
-        grantId: grantOne.id,
-        createdVia: 'rtr',
-        goalTemplateId: template.id,
-      });
-
-      goalThree = await Goal.create({
-        name: `Selected goal 3${faker.animal.dog()}`,
-        status: GOAL_STATUS.SUSPENDED,
-        endDate: null,
-        isFromSmartsheetTtaPlan: false,
-        onApprovedAR: false,
-        grantId: grantTwo.id,
-        createdVia: 'rtr',
-        goalTemplateId: template.id,
-      });
-
-      dummyGoal = await Goal.create({
-        name: dummyGoalName,
-        status: GOAL_STATUS.CLOSED,
-        endDate: null,
-        isFromSmartsheetTtaPlan: false,
-        onApprovedAR: false,
-        grantId: grantTwo.id,
-        createdVia: 'rtr',
-      });
-
-      await GoalFieldResponse.create({
-        goalTemplateFieldPromptId: prompt.id,
-        response: ['option 1', 'option 2'],
-        goalId: goalThree.id,
-      });
-
-      report = await createReport({
-        activityRecipients: [{
-          grantId: grantOne.id,
-        }],
-      });
-
-      topic = await Topic.create({
-        name: faker.datatype.string(100),
-      });
-
-      const resource = await Resource.create({
-        url: resourceUrl,
-      });
-
-      const file = await File.create({
-        originalFileName: fileName,
-        key: fileName,
-        status: FILE_STATUSES.APPROVED,
-        fileSize: 123445,
-      });
-
-      await GoalResource.create({
-        goalId: goalTwo.id,
-        resourceId: resource.id,
-      });
-
-      await ActivityReportGoal.create({
-        activityReportId: report.id,
-        goalId: goalOne.id,
-      });
-
-      await ActivityReportGoal.create({
-        activityReportId: report.id,
-        goalId: dummyGoal.id,
-      });
-
-      objectiveOneForGoalOne = await Objective.create({
-        goalId: goalOne.id,
-        title: faker.datatype.string(100),
-        status: OBJECTIVE_STATUS.NOT_STARTED,
-      });
-
-      await ActivityReportObjective.create({
-        activityReportId: report.id,
-        objectiveId: objectiveOneForGoalOne.id,
-      });
-
-      const dummyGoalObjective = await Objective.create({
-        goalId: dummyGoal.id,
-        title: faker.datatype.string(100),
-        status: OBJECTIVE_STATUS.NOT_STARTED,
-      });
-
-      await ActivityReportObjective.create({
-        activityReportId: report.id,
-        objectiveId: dummyGoalObjective.id,
-      });
-
-      const objectiveTwoForGoalOne = await Objective.create({
-        goalId: goalOne.id,
-        title: faker.datatype.string(100),
-        status: OBJECTIVE_STATUS.IN_PROGRESS,
-      });
-
-      await ObjectiveResource.create({
-        objectiveId: objectiveTwoForGoalOne.id,
-        resourceId: resource.id,
-      });
-
-      const objectiveOneForGoalTwo = await Objective.create({
-        goalId: goalTwo.id,
-        title: faker.datatype.string(100),
-        status: OBJECTIVE_STATUS.IN_PROGRESS,
-      });
-
-      await ObjectiveTopic.create({
-        objectiveId: objectiveOneForGoalTwo.id,
-        topicId: topic.id,
-      });
-
-      await ObjectiveTopic.create({
-        objectiveId: dummyGoalObjective.id,
-        topicId: topic.id,
-      });
-
-      const objectiveOneForGoalThree = await Objective.create({
-        goalId: goalThree.id,
-        title: faker.datatype.string(100),
-        status: OBJECTIVE_STATUS.IN_PROGRESS,
-      });
-
-      await ObjectiveFile.create({
-        objectiveId: objectiveOneForGoalThree.id,
-        fileId: file.id,
-      });
-
-      mergedGoals = await mergeGoals(goalOne.id, [goalOne.id, goalTwo.id, goalThree.id]);
-      mergedGoalIds = mergedGoals.map((goal) => goal.id);
-    });
-
-    afterAll(async () => {
-      const allGoals = await Goal.unscoped().findAll({
-        where: {
-          grantId: [grantOne.id, grantTwo.id, grantThree.id],
-        },
-        include: {
-          model: Objective,
-          as: 'objectives',
-        },
-      });
-
-      const allGoalIds = allGoals.map((goal) => goal.id);
-      const allObjectives = allGoals.map((g) => g.objectives).flat();
-      const allObjectiveIds = allObjectives.map((objective) => objective.id);
-
-      await ActivityReportObjective.destroy({
-        where: {
-          objectiveId: allObjectiveIds,
-        },
-      });
-
-      await ActivityReportGoal.destroy({
-        where: {
-          goalId: allGoalIds,
-        },
-      });
-
-      await ObjectiveResource.destroy({
-        where: {
-          objectiveId: allObjectiveIds,
-        },
-      });
-
-      await ObjectiveTopic.destroy({
-        where: {
-          objectiveId: allObjectiveIds,
-        },
-      });
-
-      await ObjectiveFile.destroy({
-        where: {
-          objectiveId: allObjectiveIds,
-        },
-      });
-
-      await Objective.destroy({
-        where: {
-          id: allObjectiveIds,
-        },
-        force: true,
-      });
-
-      await destroyReport(report);
-
-      await GoalFieldResponse.destroy({
-        where: {
-          goalId: allGoalIds,
-        },
-      });
-
-      await GoalResource.destroy({
-        where: {
-          goalId: allGoalIds,
-        },
-      });
-
-      await Goal.unscoped().destroy({
-        where: {
-          id: allGoalIds,
-        },
-        force: true,
-      });
-
-      await GoalTemplateFieldPrompt.destroy({
-        where: {
-          goalTemplateId: template.id,
-        },
-      });
-
-      await GoalTemplate.destroy({
-        where: {
-          id: template.id,
-        },
-      });
-
-      await Resource.destroy({
-        where: {
-          url: resourceUrl,
-        },
-      });
-
-      await File.destroy({
-        where: {
-          originalFileName: fileName,
-        },
-      });
-
-      await Topic.destroy({
-        where: {
-          name: topic.name,
-        },
-      });
-
-      await Grant.destroy({
-        where: {
-          id: [grantOne.id, grantTwo.id, grantThree.id],
-        },
-        force: true,
-      });
-
-      await Recipient.destroy({
-        where: {
-          id: recipient.id,
-        },
-      });
-    });
-
-    it('old goals are merged away', async () => {
-      expect(mergedGoals.length).toBe(2);
-
-      // verify that new goals were created
-      expect(mergedGoalIds).not.toContain(goalOne.id);
-      expect(mergedGoalIds).not.toContain(goalTwo.id);
-      expect(mergedGoalIds).not.toContain(goalThree.id);
-
-      const goalsThatAreMergedAway = await Goal.unscoped().findAll({
-        where: {
-          id: [goalOne.id, goalTwo.id, goalThree.id],
-        },
-        attributes: ['id', 'mapsToParentGoalId'],
-        include: [{
-          model: Objective,
-          as: 'objectives',
-          attributes: ['id', 'mapsToParentObjectiveId', 'goalId'],
-        }],
-      });
-
-      expect(goalsThatAreMergedAway.length).toBe(3);
-      const mapsToParentGoalIds = goalsThatAreMergedAway.map((goal) => goal.mapsToParentGoalId)
-        .sort();
-
-      const mergedGoalFromGrantOne = mergedGoals.find((g) => g.grantId === grantOne.id);
-      const otherMergedGoal = mergedGoals.find((g) => g.id !== mergedGoalFromGrantOne.id);
-
-      expect(
-        [mergedGoalFromGrantOne.id, mergedGoalFromGrantOne.id, otherMergedGoal.id].sort(),
-      ).toEqual(mapsToParentGoalIds);
-
-      const objectivesThatAreMergedAway = goalsThatAreMergedAway
-        .map((g) => g.objectives).flat();
-
-      expect(objectivesThatAreMergedAway.length).toBe(4);
-      objectivesThatAreMergedAway.forEach((objective) => {
-        expect([goalOne.id, goalTwo.id, goalThree.id]).toContain(objective.goalId);
-        expect(objective.mapsToParentObjectiveId).not.toBeNull();
-        expect(objective.mapsToParentObjectiveId).not.toBe(objective.id);
-      });
-    });
-
-    it('data is merged into new goals', async () => {
-      // verify goal & associated data with a length query
-      // Let's get everything and just go through it all
-      const goalsWithData = await Goal.findAll({
-        where: {
-          id: mergedGoalIds,
-        },
-        include: [
-          {
-            model: ActivityReportGoal,
-            as: 'activityReportGoals',
-          },
-          {
-            model: GoalFieldResponse,
-            as: 'responses',
-          },
-          {
-            model: GoalResource,
-            as: 'goalResources',
-          },
-          {
-            model: Objective,
-            as: 'objectives',
-            include: [
-              {
-                model: ObjectiveFile,
-                as: 'objectiveFiles',
-              },
-              {
-                model: ObjectiveResource,
-                as: 'objectiveResources',
-              },
-              {
-                model: ObjectiveTopic,
-                as: 'objectiveTopics',
-              },
-              {
-                model: ActivityReportObjective,
-                as: 'activityReportObjectives',
-              },
-            ],
-          },
-        ],
-        order: [['grantId', 'asc']],
-      });
-
-      expect(goalsWithData.length).toBe(2);
-      const grantIds = goalsWithData.map((goal) => goal.grantId);
-      expect(grantIds).toContain(grantOne.id);
-      expect(grantIds).toContain(grantThree.id);
-
-      const goalForGrantOne = goalsWithData.find((g) => g.grantId === grantOne.id);
-      expect(goalForGrantOne.status).toBe(GOAL_STATUS.IN_PROGRESS);
-      expect(goalForGrantOne.objectives.length).toBe(3);
-      expect(goalForGrantOne.responses.length).toBe(0);
-      expect(goalForGrantOne.goalResources.length).toBe(1);
-      expect(goalForGrantOne.activityReportGoals.length).toBe(1);
-
-      const [arGoal] = goalForGrantOne.activityReportGoals;
-      expect(arGoal.activityReportId).toBe(report.id);
-      expect(arGoal.originalGoalId).toBe(goalOne.id);
-
-      const aroForGoalForGrantOne = goalForGrantOne.objectives
-        .map((o) => o.activityReportObjectives).flat();
-      expect(aroForGoalForGrantOne.length).toBe(1);
-      expect(aroForGoalForGrantOne[0].activityReportId).toBe(report.id);
-      expect(aroForGoalForGrantOne[0].originalObjectiveId).toBe(objectiveOneForGoalOne.id);
-
-      const objectiveResourcesForGoalForGrantOne = goalForGrantOne.objectives
-        .map((o) => o.objectiveResources).flat();
-      expect(objectiveResourcesForGoalForGrantOne.length).toBe(1);
-
-      const objectiveTopicsForGoalForGrantOne = goalForGrantOne.objectives
-        .map((o) => o.objectiveTopics).flat();
-      expect(objectiveTopicsForGoalForGrantOne.length).toBe(1);
-
-      const objectiveFilesForGoalForGrantOne = goalForGrantOne.objectives
-        .map((o) => o.objectiveFiles).flat();
-      expect(objectiveFilesForGoalForGrantOne.length).toBe(0);
-
-      const goalForGrantTwo = goalsWithData.find((g) => g.grantId === grantThree.id);
-      expect(goalForGrantTwo.status).toBe(GOAL_STATUS.IN_PROGRESS);
-      expect(goalForGrantTwo.objectives.length).toBe(1);
-      expect(goalForGrantTwo.responses.length).toBe(1);
-      expect(goalForGrantTwo.goalResources.length).toBe(0);
-      expect(goalForGrantTwo.activityReportGoals.length).toBe(0);
-
-      const aroForGoalForGrantTwo = goalForGrantTwo.objectives
-        .map((o) => o.activityReportObjectives).flat();
-      expect(aroForGoalForGrantTwo.length).toBe(0);
-
-      const objectiveResourcesForGoalForGrantTwo = goalForGrantTwo.objectives
-        .map((o) => o.objectiveResources).flat();
-      expect(objectiveResourcesForGoalForGrantTwo.length).toBe(0);
-
-      const objectiveTopicsForGoalForGrantTwo = goalForGrantTwo.objectives
-        .map((o) => o.objectiveTopics).flat();
-      expect(objectiveTopicsForGoalForGrantTwo.length).toBe(0);
-
-      const objectiveFilesForGoalForGrantTwo = goalForGrantTwo.objectives
-        .map((o) => o.objectiveFiles).flat();
-      expect(objectiveFilesForGoalForGrantTwo.length).toBe(1);
-    });
-
-    it('leaves random goals and objectives alone', async () => {
-      const dummyGoalValidation = await Goal.findOne({
-        where: {
-          id: dummyGoal.id,
-        },
-        include: [
-          {
-            model: ActivityReportGoal,
-            as: 'activityReportGoals',
-          },
-          {
-            model: Objective,
-            as: 'objectives',
-            include: [
-              {
-                model: ActivityReportObjective,
-                as: 'activityReportObjectives',
-              },
-            ],
-          },
-        ],
-      });
-
-      expect(dummyGoalValidation.grantId).toBe(grantTwo.id);
-      expect(dummyGoalValidation.status).toBe(GOAL_STATUS.CLOSED);
-      expect(dummyGoalValidation.activityReportGoals.length).toBe(1);
-      expect(dummyGoalValidation.objectives.length).toBe(1);
-      expect(dummyGoalValidation.objectives[0].activityReportObjectives.length).toBe(1);
-    });
-
-    it('updates what is returned for an AR', async () => {
-      const goals = await getGoalsForReport(report.id);
-
-      expect(goals.length).toBe(2);
-
-      const goalForGrantOne = mergedGoals.find((g) => g.grantId === grantOne.id);
-      const expectedGoalIds = [goalForGrantOne.id, dummyGoal.id].sort();
-      const actualGoalIds = goals.map((goal) => goal.id).sort();
-      expect(actualGoalIds).toEqual(expectedGoalIds);
     });
   });
